requirements.txt
paho-mqtt
locust
pycryptodome
cryptography
redis
orjson
ansible